Subclassed CAMetalLayer instead of using CAMetalLayer (MoltenVK) category.
This is done only when the MVK_USE_OBJC_CLASS_CLONING macro is set.
In this case, we rely on an external #include <core_ObjCDefs.h> (that belong to Pulsar)
This will implement subclass cloning on the fly to ensure there is only one custom class with the same name in all plugins.